Move logger to different package
