+++ client.go
@@ -8,6 +8,7 @@ import (
 	"net/url"
 	"strings"
 
+	"github.com/grafana/nanogit/log"
 	"github.com/grafana/nanogit/protocol"
 	"github.com/grafana/nanogit/protocol/hash"
 	"github.com/grafana/nanogit/storage"
@@ -101,7 +102,7 @@ type Option func(*rawClient) error
 // It implements the Git Smart Protocol version 2 over HTTP/HTTPS transport.
 type httpClient struct {
 	RawClient
-	logger          Logger
+	logger          log.Logger
 	packfileStorage storage.PackfileStorage
 }
 
@@ -148,7 +149,7 @@ func NewHTTPClient(repo string, options ...Option) (Client, error) {
 	rawClient := &rawClient{
 		base:   u,
 		client: &http.Client{},
-		logger: &noopLogger{}, // No-op logger by default
+		logger: &log.NoopLogger{}, // No-op logger by default
 	}
 
 	for _, option := range options {
